divided integration and unit tests, deleted extra files
